+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Version 2022 (released XX.12.21)
 -  exclude some builds from CI (#940, #942)
 -  update interface documentation (#939)
 -  update install guide for binary packages (#936)
+-  fixed sigma plot options passing (#943)
 
 Version 2022-rc.2 (released 16.12.21)
 =====================================